Rename log.js -> console.js
