+++ packages/child/console.js
