+++ packages/parent/iife.js
@@ -1,4 +1,4 @@
-import { advise } from '../core/log'
+import { advise } from '../core/console'
 import createIframeResize from './factory'
 
 window.iframeResize = createIframeResize()
